test: replace fixed-sleep races with wait_for_condition in test_execution.py
tests/runtime/test_execution.py had a recurring anti-pattern: start
background async work, sleep a fixed duration, then assert on mock
call counts or internal state that depends on that work having
finished. Under CI load the fixed sleep isn't always long enough --
this surfaced as real flakes on PR #610's CI
(test_pending_next_message_present_in_context_still_executes, twice)
and PR #609's CI (test_sync_removes_duplicate_from_ws_queue, a
different commit, same class).

tests/runtime/test_resync.py already had the correct fix pattern
locally: a wait_for_condition(predicate, timeout, interval) poller.
Hoist it into tests/runtime/conftest.py as the shared source of
truth, and convert every settle-then-assert sleep in
test_execution.py (and two in TestExecutionContextEvents/
TestExecutionContextParticipantEvents) to wait on the exact predicate
the following assertion checks -- ctx._sync_complete, ctx.is_processing,
ctx.queue.qsize(), or a mock call/await count.

Left untouched: sleeps that are the actual behavior under test (a
deliberately slow fake handler body, a real elapsed-time threshold
assertion), an asyncio.sleep(0) yield-once idiom, and two sleeps whose
following assertions don't actually depend on the sleep's duration.

Verified with 10 repeated full-file runs, 30 repeated runs of the two
previously-observed-flaky tests, and one run under artificial CPU
contention -- zero failures, and the suite runs faster (~1.4s vs
~4-5s) since polling resolves as soon as the condition is true instead
of always waiting out the fixed sleep.
