+++ tests/runtime/test_execution.py
@@ -27,6 +27,7 @@ from tests.conftest import (
     make_participant_mock,
     make_participant_removed_event,
 )
+from tests.runtime.conftest import wait_for_condition
 
 
 @pytest.fixture
@@ -184,8 +185,7 @@ class TestExecutionContextEvents:
         event = make_message_event(room_id="room-123", msg_id="msg-1", content="Hello")
         await ctx.on_event(event)
 
-        # Wait for processing
-        await asyncio.sleep(0.1)
+        await wait_for_condition(lambda: mock_handler.call_count >= 1)
 
         mock_handler.assert_called()
         call_args = mock_handler.call_args[0]
@@ -204,9 +204,9 @@ class TestExecutionContextEvents:
 
         # Send same message twice
         await ctx.on_event(event)
-        await asyncio.sleep(0.1)
+        await wait_for_condition(lambda: mock_handler.call_count >= 1)
         await ctx.on_event(event)
-        await asyncio.sleep(0.1)
+        await wait_for_condition(lambda: ctx.queue.qsize() == 0)
 
         # Should only be called once
         assert mock_handler.call_count == 1
@@ -559,7 +559,7 @@ class TestExecutionContextParticipantEvents:
             type="User",
         )
         await ctx.on_event(event)
-        await asyncio.sleep(0.1)
+        await wait_for_condition(lambda: ctx.queue.qsize() == 0)
 
         assert any(p["id"] == "user-2" for p in ctx.participants)
 
@@ -579,7 +579,7 @@ class TestExecutionContextParticipantEvents:
             participant_id="user-1",
         )
         await ctx.on_event(event)
-        await asyncio.sleep(0.1)
+        await wait_for_condition(lambda: ctx.queue.qsize() == 0)
 
         assert not any(p["id"] == "user-1" for p in ctx.participants)
 
@@ -678,7 +678,7 @@ class TestCrashRecoverySync:
         ctx = ExecutionContext("room-123", mock_link_with_next, mock_handler)
 
         await ctx.start()
-        await asyncio.sleep(0.1)
+        await wait_for_condition(lambda: ctx._sync_complete)
 
         assert ctx._sync_complete is True
         mock_link_with_next.get_next_message.assert_called()
@@ -717,7 +717,7 @@ class TestCrashRecoverySync:
         )
 
         await ctx.start()
-        await asyncio.sleep(0.2)
+        await wait_for_condition(lambda: mock_handler.call_count >= 1)
 
         # Handler should be called for backlog message
         assert mock_handler.call_count >= 1
@@ -763,7 +763,7 @@ class TestCrashRecoverySync:
 
         # Start should sync and find sync point
         await ctx.start()
-        await asyncio.sleep(0.2)
+        await wait_for_condition(lambda: ctx._sync_complete)
 
         # Marker should be cleared
         assert ctx._first_ws_msg_id is None
@@ -818,7 +818,12 @@ class TestCrashRecoverySync:
 
         # Start triggers sync
         await ctx.start()
-        await asyncio.sleep(0.2)
+        # Two handler dispatches expected (sync message + participant event);
+        # queue must be fully drained so Phase 2's participant processing has
+        # also settled, not just the sync-point crash-recovery phase.
+        await wait_for_condition(
+            lambda: mock_handler.call_count >= 2 and ctx.queue.qsize() == 0
+        )
 
         # Sync point reached and duplicate removed from WS phase
         assert ctx._first_ws_msg_id is None
@@ -861,7 +866,7 @@ class TestCrashRecoverySync:
             config=SessionConfig(enable_context_hydration=False),
         )
         await ctx.start()
-        await asyncio.sleep(0.05)
+        await wait_for_condition(lambda: ctx._sync_complete)
 
         event = make_message_event(
             room_id="room-123",
@@ -872,7 +877,7 @@ class TestCrashRecoverySync:
             ),
         )
         await ctx.on_event(event)
-        await asyncio.sleep(0.1)
+        await wait_for_condition(lambda: ctx.queue.qsize() == 0)
 
         mock_handler.assert_not_called()
         mock_link_with_next.mark_processing.assert_not_called()
@@ -907,11 +912,11 @@ class TestCrashRecoverySync:
             config=SessionConfig(enable_context_hydration=True),
         )
         await ctx.start()
-        await asyncio.sleep(0.05)
+        await wait_for_condition(lambda: ctx._sync_complete)
 
         event = make_message_event(room_id="room-123", msg_id="msg-stale-replay")
         await ctx.on_event(event)
-        await asyncio.sleep(0.1)
+        await wait_for_condition(lambda: ctx.queue.qsize() == 0)
 
         mock_handler.assert_not_called()
         mock_link_with_next.mark_processing.assert_not_called()
@@ -964,7 +969,9 @@ class TestCrashRecoverySync:
         )
 
         await ctx.start()
-        await asyncio.sleep(0.2)
+        await wait_for_condition(
+            lambda: mock_link_with_next.mark_processed.call_count >= 1
+        )
 
         mock_link_with_next.mark_processing.assert_called_once_with(
             "room-123", "msg-pending-down"
@@ -1077,7 +1084,10 @@ class TestCrashRecoverySync:
         await ctx.on_event(make_message_event(room_id="room-123", msg_id="msg-first"))
 
         release_handler.set()
-        await asyncio.sleep(0.2)  # Let sync finish and Phase 2 drain the WS copy
+        # Let sync finish and Phase 2 drain the WS copy.
+        await wait_for_condition(
+            lambda: mock_link_with_next.mark_processed.await_count >= 1
+        )
 
         assert handled_message_ids == ["msg-first"]
         assert mock_link_with_next.mark_processing.await_count == 1
@@ -1399,10 +1409,12 @@ class TestCrashRecoverySync:
         )
 
         await ctx.start()
-        await asyncio.sleep(0.1)
+        await wait_for_condition(lambda: ctx._sync_complete)
         await ctx.on_event(make_message_event(room_id="room-123", msg_id="msg-old-ws"))
         await ctx.on_event(make_message_event(room_id="room-123", msg_id="msg-new-ws"))
-        await asyncio.sleep(0.3)
+        await wait_for_condition(
+            lambda: mock_link_with_next.mark_processed.await_count >= 3
+        )
 
         assert [call.args[1].payload.id for call in mock_handler.await_args_list] == [
             "msg-old-ws",
@@ -1491,7 +1503,9 @@ class TestCrashRecoverySync:
             make_message_event(room_id="room-123", msg_id="msg-sync-claim-fails")
         )
         await ctx.start()
-        await asyncio.sleep(0.2)
+        await wait_for_condition(
+            lambda: mock_link_with_next.mark_processing.await_count >= 1
+        )
 
         assert ctx._first_ws_msg_id == "msg-sync-claim-fails"
         mock_handler.assert_not_called()
@@ -1569,7 +1583,9 @@ class TestCrashRecoverySync:
             make_message_event(room_id="room-123", msg_id="msg-newer-ws")
         )
         await ctx.start()
-        await asyncio.sleep(0.2)
+        await wait_for_condition(
+            lambda: mock_link_with_next.mark_processing.await_count >= 1
+        )
 
         mock_link_with_next.mark_processing.assert_awaited_once_with(
             "room-123", "msg-older-claim-fails"
@@ -1646,12 +1662,14 @@ class TestCrashRecoverySync:
         )
 
         await ctx.start()
-        await asyncio.sleep(0.1)
+        await wait_for_condition(lambda: ctx._sync_complete)
         await ctx.request_resync()
         await ctx.on_event(
             make_message_event(room_id="room-123", msg_id="msg-resync-newer-ws")
         )
-        await asyncio.sleep(0.2)
+        await wait_for_condition(
+            lambda: mock_link_with_next.mark_processing.await_count >= 1
+        )
 
         mock_link_with_next.mark_processing.assert_awaited_once_with(
             "room-123", "msg-resync-older-claim-fails"
@@ -1693,7 +1711,7 @@ class TestCrashRecoverySync:
         ctx._retry_tracker.mark_permanently_failed("msg-failed-001")
 
         await ctx.start()
-        await asyncio.sleep(0.2)
+        await wait_for_condition(lambda: ctx._sync_complete)
 
         # Handler should NOT be called for failed message
         assert mock_handler.call_count == 0
@@ -1860,14 +1878,14 @@ class TestCancellationDuringProcessing:
         await ctx.start()
 
         # Wait for sync to complete
-        await asyncio.sleep(0.05)
+        await wait_for_condition(lambda: ctx._sync_complete)
 
         # Enqueue a message to trigger processing
         event = make_message_event(room_id="room-123", msg_id="msg-001", content="Test")
         await ctx.on_event(event)
 
-        # Give time to start processing
-        await asyncio.sleep(0.05)
+        # Wait for the slow handler to actually be in flight
+        await wait_for_condition(lambda: ctx.is_processing)
 
         # Stop should cancel processing
         start = asyncio.get_running_loop().time()
@@ -2093,11 +2111,11 @@ class TestContextCacheTTL:
         ctx._context_hydrated = True
 
         await ctx.start()
-        await asyncio.sleep(0.05)
+        await wait_for_condition(lambda: ctx._sync_complete)
 
         event = make_message_event(room_id="room-123", msg_id="msg-ttl")
         await ctx.on_event(event)
-        await asyncio.sleep(0.1)
+        await wait_for_condition(lambda: mock_handler.call_count >= 1)
 
         mock_handler.assert_called()
         assert mock_link.rest.agent_api_context.get_agent_chat_context.await_count == 1
@@ -2126,7 +2144,7 @@ class TestParticipantCallbacks:
             on_participant_added=on_participant_added,
         )
         await ctx.start()
-        await asyncio.sleep(0.05)
+        await wait_for_condition(lambda: ctx._sync_complete)
 
         event = make_participant_added_event(
             room_id="room-123",
@@ -2134,7 +2152,7 @@ class TestParticipantCallbacks:
             name="User Two",
         )
         await ctx.on_event(event)
-        await asyncio.sleep(0.1)
+        await wait_for_condition(lambda: mock_handler.await_count >= 1)
 
         on_participant_added.assert_awaited_once_with("room-123", event)
         mock_handler.assert_awaited_once()
@@ -2158,14 +2176,14 @@ class TestParticipantCallbacks:
             on_participant_removed=on_participant_removed,
         )
         await ctx.start()
-        await asyncio.sleep(0.05)
+        await wait_for_condition(lambda: ctx._sync_complete)
 
         event = make_participant_removed_event(
             room_id="room-123",
             participant_id="user-1",
         )
         await ctx.on_event(event)
-        await asyncio.sleep(0.1)
+        await wait_for_condition(lambda: mock_handler.await_count >= 1)
 
         on_participant_removed.assert_awaited_once_with("room-123", event)
         mock_handler.assert_awaited_once()
@@ -2185,7 +2203,7 @@ class TestParticipantCallbacks:
             on_participant_added=on_participant_added,
         )
         await ctx.start()
-        await asyncio.sleep(0.05)
+        await wait_for_condition(lambda: ctx._sync_complete)
 
         event = make_participant_added_event(
             room_id="room-123",
@@ -2193,7 +2211,7 @@ class TestParticipantCallbacks:
             name="User Two",
         )
         await ctx.on_event(event)
-        await asyncio.sleep(0.1)
+        await wait_for_condition(lambda: mock_handler.await_count >= 1)
 
         on_participant_added.assert_awaited_once()
         mock_handler.assert_awaited_once()
@@ -2272,14 +2290,14 @@ class TestGracefulStopWithTimeout:
             config=SessionConfig(enable_context_hydration=False),
         )
         await ctx.start()
-        await asyncio.sleep(0.05)
+        await wait_for_condition(lambda: ctx._sync_complete)
 
         # Enqueue a message
         event = make_message_event(room_id="room-123", msg_id="msg-001")
         await ctx.on_event(event)
 
-        # Give time to start processing
-        await asyncio.sleep(0.05)
+        # Wait for the handler to actually be in flight
+        await wait_for_condition(lambda: ctx.is_processing)
 
         # Stop with timeout - should wait for processing
         result = await ctx.stop(timeout=5.0)
@@ -2300,14 +2318,14 @@ class TestGracefulStopWithTimeout:
             config=SessionConfig(enable_context_hydration=False),
         )
         await ctx.start()
-        await asyncio.sleep(0.05)
+        await wait_for_condition(lambda: ctx._sync_complete)
 
         # Enqueue a message
         event = make_message_event(room_id="room-123", msg_id="msg-001")
         await ctx.on_event(event)
 
-        # Give time to start processing
-        await asyncio.sleep(0.05)
+        # Wait for the handler to actually be in flight
+        await wait_for_condition(lambda: ctx.is_processing)
 
         # Stop with short timeout
         start = asyncio.get_running_loop().time()
